+++ src/main/java/courgette/runtime/CourgetteNoObjectFactory.java
@@ -1,0 +1,23 @@
+package courgette.runtime;
+
+import io.cucumber.core.backend.ObjectFactory;
+
+public class CourgetteNoObjectFactory implements ObjectFactory {
+    @Override
+    public void start() {
+    }
+
+    @Override
+    public void stop() {
+    }
+
+    @Override
+    public boolean addClass(Class<?> glueClass) {
+        return false;
+    }
+
+    @Override
+    public <T> T getInstance(Class<T> glueClass) {
+        return null;
+    }
+}