+++ src/main/java/courgette/runtime/CourgettePickleMatcher.java
@@ -1,22 +1,18 @@
 package courgette.runtime;
 
-import cucumber.runtime.filter.Filters;
-import cucumber.runtime.model.CucumberFeature;
-import gherkin.events.PickleEvent;
-import gherkin.pickles.Compiler;
-import gherkin.pickles.Pickle;
 import gherkin.pickles.PickleLocation;
+import io.cucumber.core.feature.CucumberFeature;
+import io.cucumber.core.feature.CucumberPickle;
+import io.cucumber.core.filter.Filters;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 public class CourgettePickleMatcher {
-    private final Compiler compiler;
     private final CucumberFeature cucumberFeature;
     private final Filters filters;
 
     public CourgettePickleMatcher(CucumberFeature cucumberFeature, Filters filters) {
-        this.compiler = new Compiler();
         this.cucumberFeature = cucumberFeature;
         this.filters = filters;
     }
@@ -25,38 +21,33 @@ public class CourgettePickleMatcher {
         AtomicBoolean matched = new AtomicBoolean();
 
         try {
-            compiler.compile(cucumberFeature.getGherkinFeature()).forEach(pickle -> {
-                matched.set(filters.matchesFilters(new PickleEvent(cucumberFeature.getUri().getSchemeSpecificPart(), pickle)));
-
+            cucumberFeature.getPickles().forEach(pickle -> {
+                matched.set(filters.test(pickle));
                 if (matched.get()) {
                     throw new ConditionSatisfiedException();
                 }
             });
         } catch (ConditionSatisfiedException ignored) {
         }
-
         return matched.get();
     }
 
     public PickleLocation matchLocation(int pickleLocationLine) {
         final PickleLocation[] location = {null};
 
-        List<Pickle> pickles = compiler.compile(cucumberFeature.getGherkinFeature());
+        List<CucumberPickle> pickles = cucumberFeature.getPickles();
 
         try {
-            pickles.forEach(pickle -> {
-                PickleLocation pickleLocation = pickle.getLocations().stream().filter(l -> l.getLine() == pickleLocationLine).findFirst().orElse(null);
-
-                if (pickleLocation != null) {
-                    if (filters.matchesFilters(new PickleEvent(cucumberFeature.getUri().getSchemeSpecificPart(), pickle))) {
-                        location[0] = pickleLocation;
-                        throw new ConditionSatisfiedException();
-                    }
-                }
-            });
+            pickles.stream().filter(p -> p.getLine() == pickleLocationLine)
+                    .findFirst()
+                    .ifPresent(pickleEvent -> {
+                        if (filters.test(pickleEvent)) {
+                            location[0] = new PickleLocation(pickleEvent.getLine(), pickleEvent.getScenarioLine());
+                            throw new ConditionSatisfiedException();
+                        }
+                    });
         } catch (ConditionSatisfiedException ignored) {
         }
-
         return location[0];
     }