+++ src/main/java/courgette/runtime/CourgetteRunOptions.java
@@ -2,9 +2,12 @@ package courgette.runtime;
 
 import courgette.api.CourgetteOptions;
 import courgette.api.CourgetteRunLevel;
+import courgette.api.CucumberOptions;
+import courgette.integration.reportportal.ReportPortalProperties;
+import courgette.runtime.utils.FileUtils;
 import courgette.runtime.utils.SystemPropertyUtils;
-import cucumber.api.CucumberOptions;
 
+import java.io.File;
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
 
@@ -13,6 +16,7 @@ public class CourgetteRunOptions implements CourgetteOptions {
 
     public CourgetteRunOptions(Class clazz) {
         validate(clazz);
+        validatePlugins();
     }
 
     @Override
@@ -51,6 +55,11 @@ public class CourgetteRunOptions implements CourgetteOptions {
     }
 
     @Override
+    public String[] plugin() {
+        return courgetteOptions.plugin();
+    }
+
+    @Override
     public Class<? extends Annotation> annotationType() {
         return null;
     }
@@ -60,5 +69,24 @@ public class CourgetteRunOptions implements CourgetteOptions {
                 .filter(annotation -> annotation.annotationType().equals(CourgetteOptions.class))
                 .findFirst()
                 .orElseThrow(() -> new CourgetteException("Runner class is not annotated with @CourgetteOptions"));
+    }
+
+    private void validatePlugins() {
+        if (plugin().length > 0) {
+            validateReportPortalPlugin();
+        }
+    }
+
+    private void validateReportPortalPlugin() {
+        final String reportPortalPropertiesFilename = "reportportal.properties";
+
+        if (Arrays.stream(courgetteOptions.plugin()).anyMatch(plugin -> plugin.equalsIgnoreCase("reportportal"))) {
+            File reportPortalPropertiesFile = FileUtils.getClassPathFile(reportPortalPropertiesFilename);
+
+            if (reportPortalPropertiesFile == null) {
+                throw new CourgetteException("The " + reportPortalPropertiesFilename + " file must be in your classpath to use the Courgette reportportal plugin");
+            }
+            new ReportPortalProperties().validate();
+        }
     }
 }