+++ src/main/java/courgette/runtime/CourgetteRunResult.java
@@ -1,12 +1,26 @@
 package courgette.runtime;
 
+import io.cucumber.core.feature.CucumberFeature;
+
 public class CourgetteRunResult {
+    private CucumberFeature cucumberFeature;
+    private Integer lineId;
     private String featureUri;
     private Status status;
 
-    public CourgetteRunResult(String featureUri, Status status) {
+    public CourgetteRunResult(CucumberFeature cucumberFeature, Integer lineId, String featureUri, Status status) {
+        this.cucumberFeature = cucumberFeature;
+        this.lineId = lineId;
         this.featureUri = featureUri;
         this.status = status;
+    }
+
+    public CucumberFeature getCucumberFeature() {
+        return cucumberFeature;
+    }
+
+    public Integer getLineId() {
+        return lineId;
     }
 
     public String getFeatureUri() {